simple: Fix deadlock due to slow data mutex release
In the Drop instance for our MutexGuard, we need to drop the inner guard
before waking the next thread, to ensure that the next thread doesn't
immediately go back to sleep again with noone to wake it.  Unfortunately
struct fields are dropped *after* the struct's destructor is run.  This
patch is essentially a work-around.